refactor: split extract command into submodules
Extract from `cmd/extract_cmd.rs` (1880 lines):
- `cmd/extract_cmd/config.rs` (198 lines) — ImportersFile, ImporterEntry,
  config loading, importer matching
- `cmd/extract_cmd/duplicate.rs` (90 lines) — duplicate detection,
  fuzzy text matching, OFX detection

mod.rs retains Args, main entry point, run(), and tests.

No behavioral changes.
